upgrade rspec and separating spec helper and rails helper
