+++ spec/jobs/change_request_job_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 require 'sucker_punch/testing/inline'
 
 describe ChangeRequestJob, :type => :job do
