+++ spec/jobs/incident_report_job_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 require 'sucker_punch/testing/inline'
 
 describe IncidentReportJob, :type => :job do
