+++ spec/jobs/incident_report_new_slack_notif_job_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 require 'sucker_punch/testing/inline'
 require 'webmock/rspec'
