+++ spec/jobs/new_change_request_slack_notification_job_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 require 'sucker_punch/testing/inline'
 
 RSpec.describe NewChangeRequestSlackNotificationJob, type: :job do
