+++ spec/lib/csv_exporter.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 require 'csv_exporter'
 
 describe CSVExporter do
