+++ spec/lib/jira_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 require 'jira'
 require 'jira_helpers.rb'
