+++ spec/lib/notifier_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 require 'notifier'
 
 describe Notifier do
