+++ spec/lib/slack_attachment_builder_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 require 'slack_attachment_builder.rb'
 
 describe SlackAttachmentBuilder do
