+++ spec/models/access_request_comment_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 RSpec.describe AccessRequestComment, type: :model do
