+++ spec/models/change_request_status_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ChangeRequestStatus, type: :model do
