+++ spec/models/comment_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe Comment, type: :model do
   #shoulda matchers test
