+++ spec/models/incident_report_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe IncidentReport, type: :model do
