+++ spec/parser/access_request_csv_parser_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 require 'mentioner'
 
 describe AccessRequestCsvParser do
