+++ spec/spec_helper.rb
@@ -1,16 +1,5 @@
-# This file is copied to spec/ when you run 'rails generate rspec:install'
-require 'dotenv'
-Dotenv.load('.test.env')
-require 'capybara-screenshot/rspec'
 require 'simplecov'
 SimpleCov.start 'rails'
-require File.expand_path("../../config/environment", __FILE__)
-require 'rspec/rails'
-require 'webmock/rspec'
-require 'sunspot/rails/spec_helper'
-require 'capybara/rspec'
-require 'capybara/rails'
-require 'sucker_punch/testing/inline'
 
 SimpleCov.at_exit do
   SimpleCov.result.format!
@@ -26,142 +15,102 @@ SimpleCov.at_exit do
   end
 end
 
-WebMock.disable_net_connect!(allow_localhost: true)
-
-# Requires supporting ruby files with custom matchers and macros, etc,
-# in spec/support/ and its subdirectories.
-Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
-
-# Checks for pending migrations before tests are run.
-# If you are not using ActiveRecord, you can remove this line.
-ActiveRecord::Migration.maintain_test_schema!
-
+# This file was generated by the `rails generate rspec:install` command. Conventionally, all
+# specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
+# The generated `.rspec` file contains `--require spec_helper` which will cause
+# this file to always be loaded, without a need to explicitly require it in any
+# files.
+#
+# Given that it is always loaded, you are encouraged to keep this file as
+# light-weight as possible. Requiring heavyweight dependencies from this file
+# will add to the boot time of your test suite on EVERY test run, even for an
+# individual file that may not need all of that loaded. Instead, consider making
+# a separate helper file that requires the additional dependencies and performs
+# the additional setup, and require it from the spec files that actually need
+# it.
+#
+# The `.rspec` file also contains a few flags that are not defaults but that
+# users commonly want.
+#
+# See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
-  config.before(:each) do
-    ::Sunspot.session = ::Sunspot::Rails::StubSessionProxy.new(::Sunspot.session)
-
-    cal_response_json = File.read(File.expand_path("../webmocks/calendar_response.json", __FILE__))
-    cal_add_response_json = File.read(File.expand_path("../webmocks/add_calendar.json", __FILE__))
-    get_cal_response_json = File.read(File.expand_path("../webmocks/get_calendar.json", __FILE__))
-
-    # get calendar
-    stub_request(:get, "https://www.googleapis.com/discovery/v1/apis/calendar/v3/rest")
-      .with(headers: {'Accept'=>'*/*'})
-      .to_return(status: 200, body: cal_response_json, headers: {
-        'Content-location' => 'https://www.googleapis.com/discovery/v1/apis/calendar/v3/rest',
-        'Server' => 'GSE',
-        'Content-type' => 'application/json',
-        'charset' => 'UTF-8'
-    })
-
-    # add to calendar
-    stub_request(:post, "https://www.googleapis.com/calendar/v3/calendars/primary/events?sendNotifications=true")
-      .with(headers: {'Accept'=>'*/*', 'Accept-Encoding'=>'gzip',
-                      'Authorization'=>'Bearer 123456', 'Cache-Control'=>'no-store',
-                      'Content-Type'=>'application/json'})
-      .to_return(status: 200, body: cal_add_response_json, headers: {
-        'Server' => 'GSE',
-        'Content-type' => 'application/json'
-    })
-
-    stub_request(:post, "https://www.googleapis.com/calendar/v3/calendars//events?sendNotifications=true")
-      .to_return(status: 200, body: cal_add_response_json, headers: {
-        'Server' => 'GSE',
-        'Content-type' => 'application/json'
-      })
-    # Event calender
-    stub_request(:post, "https://www.googleapis.com/calendar/v3/calendars/veritrans.co.id_u8h6tgnhgedrt0c2ognpe7q3q0@group.calendar.google.com/events?sendNotifications=true")
-      .to_return(status: 403, body: JSON.dump({
-        "error":{
-          "errors":[{
-            "domain":"calendar",
-            "reason":"requiredAccessLevel",
-            "message":"You need to have writer access to this calendar."}],
-            "code":403,"message":"You need to have writer access to this calendar."}}))
-
-    # delete calendar
-    stub_request(:delete, "https://www.googleapis.com/calendar/v3/calendars/primary/events/?sendNotifications=true").
-      with(:headers => {'Accept'=>'*/*', 'Authorization'=>'Bearer 123456'})
-      .to_return(:status => 200, :body => "", :headers => {})
-
-    # get/read calendar
-    stub_request(:get, "https://www.googleapis.com/calendar/v3/calendars/primary/events/")
-      .with(headers: {'Accept'=>'*/*', 'Authorization'=>'Bearer 123456'})
-      .to_return(status: 200, body: get_cal_response_json, headers: {
-        'Server' => 'GSE',
-        'Content-type' => 'application/json'
-    })
-
-    # put calendar
-    stub_request(:put, "https://www.googleapis.com/calendar/v3/calendars/primary/events/?sendNotifications=true")
-      .with(headers: {'Accept'=>'*/*', 'Authorization'=>'Bearer 123456'})
-      .to_return(status: 200, body: "", headers: {})
-
-    # slack notification
-    stub_request(:post, "https://slack.com/api/chat.postMessage")
-      .to_return(status: 200, body: '{"ok": true}', headers: {})
-
-    stub_request(:post, "https://slack.com/api/users.list")
-      .to_return(status: 200, body: '{"ok": true}', headers: {})
-
-    # clear action mailer
-    ActionMailer::Base.deliveries = []
-
+  # rspec-expectations config goes here. You can use an alternate
+  # assertion/expectation library such as wrong or the stdlib/minitest
+  # assertions if you prefer.
+  config.expect_with :rspec do |expectations|
+    # This option will default to `true` in RSpec 4. It makes the `description`
+    # and `failure_message` of custom matchers include text for helper methods
+    # defined using `chain`, e.g.:
+    #     be_bigger_than(2).and_smaller_than(4).description
+    #     # => "be bigger than 2 and smaller than 4"
+    # ...rather than:
+    #     # => "be bigger than 2"
+    expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
 
-  config.after(:each) do
-    ::Sunspot.session = ::Sunspot.session.original_session
+  # rspec-mocks config goes here. You can use an alternate test double
+  # library (such as bogus or mocha) by changing the `mock_with` option here.
+  config.mock_with :rspec do |mocks|
+    # Prevents you from mocking or stubbing a method that does not exist on
+    # a real object. This is generally recommended, and will default to
+    # `true` in RSpec 4.
+    mocks.verify_partial_doubles = true
   end
 
-  # ## Mock Framework
-  #
-  # If you prefer to use mocha, flexmock or RR, uncomment the appropriate line:
-  #
-  # config.mock_with :mocha
-  # config.mock_with :flexmock
-  # config.mock_with :rr
+  # This option will default to `:apply_to_host_groups` in RSpec 4 (and will
+  # have no way to turn it off -- the option exists only for backwards
+  # compatibility in RSpec 3). It causes shared context metadata to be
+  # inherited by the metadata hash of host groups and examples, rather than
+  # triggering implicit auto-inclusion in groups with matching metadata.
+  config.shared_context_metadata_behavior = :apply_to_host_groups
 
-  # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
-  config.fixture_path = "#{::Rails.root}/spec/fixtures"
+# The settings below are suggested to provide a good initial experience
+# with RSpec, but feel free to customize to your heart's content.
+=begin
+  # This allows you to limit a spec run to individual examples or groups
+  # you care about by tagging them with `:focus` metadata. When nothing
+  # is tagged with `:focus`, all examples get run. RSpec also provides
+  # aliases for `it`, `describe`, and `context` that include `:focus`
+  # metadata: `fit`, `fdescribe` and `fcontext`, respectively.
+  config.filter_run_when_matching :focus
 
-  # If you're not using ActiveRecord, or you'd prefer not to run each of your
-  # examples within a transaction, remove the following line or assign false
-  # instead of true.
-  config.use_transactional_fixtures = true
+  # Allows RSpec to persist some state between runs in order to support
+  # the `--only-failures` and `--next-failure` CLI options. We recommend
+  # you configure your source control system to ignore this file.
+  config.example_status_persistence_file_path = "spec/examples.txt"
 
-  # If true, the base class of anonymous controllers will be inferred
-  # automatically. This will be the default behavior in future versions of
-  # rspec-rails.
-  config.infer_base_class_for_anonymous_controllers = false
+  # Limits the available syntax to the non-monkey patched syntax that is
+  # recommended. For more details, see:
+  #   - http://rspec.info/blog/2012/06/rspecs-new-expectation-syntax/
+  #   - http://www.teaisaweso.me/blog/2013/05/27/rspecs-new-message-expectation-syntax/
+  #   - http://rspec.info/blog/2014/05/notable-changes-in-rspec-3/#zero-monkey-patching-mode
+  config.disable_monkey_patching!
+
+  # Many RSpec users commonly either run the entire suite or an individual
+  # file, and it's useful to allow more verbose output when running an
+  # individual spec file.
+  if config.files_to_run.one?
+    # Use the documentation formatter for detailed output,
+    # unless a formatter has already been configured
+    # (e.g. via a command-line flag).
+    config.default_formatter = 'doc'
+  end
+
+  # Print the 10 slowest examples and example groups at the
+  # end of the spec run, to help surface which specs are running
+  # particularly slow.
+  config.profile_examples = 10
 
   # Run specs in random order to surface order dependencies. If you find an
   # order dependency and want to debug it, you can fix the order by providing
   # the seed, which is printed after each run.
   #     --seed 1234
-  config.order = "random"
+  config.order = :random
 
-  config.include Warden::Test::Helpers
-  config.include Devise::Test::ControllerHelpers, :type => :controller
-  config.infer_spec_type_from_file_location!
-
-end
-Shoulda::Matchers.configure do |config|
-  config.integrate do |with|
-    with.test_framework :rspec
-    with.library :rails
-  end
-end
-
-def login_as(user)
-  OmniAuth.config.test_mode = true
-  OmniAuth.config.add_mock(:google_oauth2, {
-    uid: user.uid,
-    credentials: {
-      token: "token",
-      refresh_token: "refresh_token",
-      expired_at: Time.now + 7.days
-    }
-  })
-  visit("/users/auth/google_oauth2")
-  OmniAuth.config.test_mode = false
+  # Seed global randomization in this process using the `--seed` CLI option.
+  # Setting this allows you to use `--seed` to deterministically reproduce
+  # test failures related to randomization by passing the same `--seed` value
+  # as the one that triggered the failure.
+  Kernel.srand config.seed
+=end
 end
